Refactor auth routes into separate file
